Add new CRUD methods to Club API
- Added setters (they do not update the backend)
- Added update methods that apply setters and save changes in backend

Add new CRUD methods to Club API

- Added setters (they do not update the backend)
- Added update methods that apply setters and save changes in backend
