+++ app/src/main/AndroidManifest.xml
@@ -16,21 +16,20 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.SquadHouse"
         android:usesCleartextTraffic="true">
-        <activity android:name=".activities.SignUpActivity"></activity>
         <activity android:name=".activities.ClubActivity"></activity>
-        <activity android:name=".activities.ViewMyProfileActivity" />
-        <activity android:name=".activities.SettingsActivity" />
+        <activity android:name=".activities.ProfileActivity"></activity>
+        <activity android:name=".activities.SettingsActivity"></activity>
         <activity android:name=".activities.LoginActivity">
+
             <!-- Set default startup screen -->
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-
         </activity>
-        <activity android:name=".activities.HomeActivity" />
-        <activity android:name=".activities.WelcomeActivity" />
+        <activity android:name=".activities.HomeActivity"></activity>
+        <activity android:name=".activities.WelcomeActivity"></activity>
     </application>
 
 </manifest>
