+++ app/src/main/java/com/pepetech/squadhouse/activities/LoginActivity.java
@@ -80,13 +80,12 @@ public class LoginActivity extends AppCompatActivity {
         Log.i(TAG, "Attempting to sign up user");
         Intent i = new Intent(this, SignUpActivity.class);
         startActivity(i);
-        overridePendingTransition(R.anim.slide_from_right, R.anim.slide_to_left);
+        finish(); // disable user ability to renavigate after a successful login
     }
 
     private void goToHomeActivity() {
         Intent i = new Intent(this, HomeActivity.class);
         startActivity(i);
-        overridePendingTransition(R.anim.slide_from_right, R.anim.slide_to_left);
         finish(); // disable user ability to renavigate after a successful login
     }
 }
